requirements.txt
appnope==0.1.4
astroid==3.3.5
asttokens==2.4.1
boto3==1.35.39
botocore==1.35.39
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
dill==0.3.9
executing==2.1.0
fonttools==4.54.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
isort==5.13.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pylint==3.3.1
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
s3transfer==0.10.3
six==1.16.0
stack-data==0.6.3
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
confluent_kafka